requirements.txt
python-telegram-bot==22.5
apscheduler==3.11.0
openai==2.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
sympy==1.12
requests==2.31.0
